+++ compositor_render/src/renderer/texture.rs
@@ -1,10 +1,15 @@
-use std::sync::{Arc, Mutex};
+use std::{
+    io::Write,
+    sync::{Arc, Mutex},
+};
 
-use bytes::Bytes;
+use bytes::{BufMut, Bytes, BytesMut};
 use compositor_common::{scene::Resolution, util::RGBColor, Frame};
-use wgpu::BufferAsyncError;
+use crossbeam_channel::bounded;
+use log::error;
+use wgpu::{Buffer, BufferAsyncError, MapMode};
 
-use self::yuv::YUVPendingDownload;
+use self::{utils::pad_to_256, yuv::YUVPendingDownload};
 
 use super::WgpuCtx;
 
@@ -162,9 +167,37 @@ impl OutputTexture {
         self.textures.copy_to_buffers(ctx, &self.buffers);
 
         YUVPendingDownload::new(
-            self.textures.planes[0].prepare_download(&self.buffers[0]),
-            self.textures.planes[1].prepare_download(&self.buffers[1]),
-            self.textures.planes[2].prepare_download(&self.buffers[2]),
+            self.download_buffer(self.textures.planes[0].texture.size(), &self.buffers[0]),
+            self.download_buffer(self.textures.planes[1].texture.size(), &self.buffers[1]),
+            self.download_buffer(self.textures.planes[2].texture.size(), &self.buffers[2]),
         )
+    }
+
+    fn download_buffer<'a>(
+        &'a self,
+        size: wgpu::Extent3d,
+        source: &'a Buffer,
+    ) -> impl FnOnce() -> Result<Bytes, BufferAsyncError> + 'a {
+        let buffer = BytesMut::with_capacity((size.width * size.height) as usize);
+        let (s, r) = bounded(1);
+        source.slice(..).map_async(MapMode::Read, move |result| {
+            if let Err(err) = s.send(result) {
+                error!("channel send error: {err}")
+            }
+        });
+
+        move || {
+            r.recv().unwrap()?;
+            let mut buffer = buffer.writer();
+            {
+                let range = source.slice(..).get_mapped_range();
+                let chunks = range.chunks(pad_to_256(size.width) as usize);
+                for chunk in chunks {
+                    buffer.write_all(&chunk[..size.width as usize]).unwrap();
+                }
+            };
+            source.unmap();
+            Ok(buffer.into_inner().into())
+        }
     }
 }
